+++ ChemoTracker/components/Calendar.js
@@ -1,7 +1,8 @@
 import React, { Component } from 'react';
-import { Text, View, Button, StyleSheet } from 'react-native';
+import { Text, View } from 'react-native';
 import { TabNavigator } from 'react-navigation';
 import Icon from 'react-native-vector-icons/FontAwesome';
+import styles from '../styles/main.js';
 
 class Calendar extends Component {
   static navigationOptions = {
@@ -17,14 +18,5 @@ class Calendar extends Component {
     );
   }
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
 
 export default Calendar;
