+++ ChemoTracker/components/Search.js
@@ -1,8 +1,9 @@
 import React, { Component } from 'react';
-import { View, Text, StyleSheet } from 'react-native';
+import { View, Text } from 'react-native';
 import { TabNavigator } from 'react-navigation';
 import Navigation from './Navigation.js';
 import Icon from 'react-native-vector-icons/FontAwesome';
+import styles from '../styles/main.js';
 
 class Search extends Component {
   static navigationOptions = {
@@ -18,14 +19,5 @@ class Search extends Component {
     );
   }
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
 
 export default Search;
