+++ ChemoTracker/styles/main.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { StyleSheet } from 'react-native';
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
+
+export default styles;
